+++ CinkesCollision/CEPA.h
@@ -37,7 +37,7 @@ namespace Cinkes
 		void BlowUp(CSimplex& a_Simplex, const std::shared_ptr<CContactInfo>& a_Contact);
 		CVector3 CSOSupport(const CContactInfo* a_Contact, const CVector3& a_Dir);
 		CVector3 SmallestAxis(const CVector3& a_Vector3);
-		std::pair<std::vector<CFaceData>, unsigned> GetFaceNormals(const std::vector<CVector3>& a_Polytope, const std::vector<unsigned>& a_Faces);
-		void AddUniqueEdge(std::vector<std::pair<unsigned, unsigned>>& a_Edges, const std::vector<unsigned>& a_Faces, unsigned a_A, unsigned a_B);
+		std::pair<std::vector<CFaceData>, size_t> GetFaceNormals(const std::vector<CVector3>& a_Polytope, const std::vector<size_t>& a_Faces);
+		void AddUniqueEdge(std::vector<std::pair<size_t, size_t>>& a_Edges, const std::vector<size_t>& a_Faces, size_t a_A, size_t a_B);
 	};
 }
